ibc_app_transfer::types::proto::transfer::v1::query_server
Trait Query

pub trait Query:
    Send
    + Sync
    + 'static {
    // Required methods
    fn denom_traces<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDenomTracesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomTracesResponse>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn denom_trace<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDenomTraceRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomTraceResponse>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn params<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryParamsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryParamsResponse>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn denom_hash<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDenomHashRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomHashResponse>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn escrow_address<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryEscrowAddressRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryEscrowAddressResponse>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
    fn total_escrow_for_denom<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryTotalEscrowForDenomRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTotalEscrowForDenomResponse>, Status>> + Send + 'async_trait>>
       where 'life0: 'async_trait,
             Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with QueryServer.

Required Methods

fn denom_traces<'life0, 'async_trait>( &'life0 self, request: Request<QueryDenomTracesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomTracesResponse>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

DenomTraces queries all denomination traces.

fn denom_trace<'life0, 'async_trait>( &'life0 self, request: Request<QueryDenomTraceRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomTraceResponse>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

DenomTrace queries a denomination trace information.

fn params<'life0, 'async_trait>( &'life0 self, request: Request<QueryParamsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryParamsResponse>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

Params queries all parameters of the ibc-transfer module.

fn denom_hash<'life0, 'async_trait>( &'life0 self, request: Request<QueryDenomHashRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryDenomHashResponse>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

DenomHash queries a denomination hash information.

fn escrow_address<'life0, 'async_trait>( &'life0 self, request: Request<QueryEscrowAddressRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryEscrowAddressResponse>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

EscrowAddress returns the escrow address for a particular port and channel id.

fn total_escrow_for_denom<'life0, 'async_trait>( &'life0 self, request: Request<QueryTotalEscrowForDenomRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTotalEscrowForDenomResponse>, Status>> + Send + 'async_trait>>
where 'life0: 'async_trait, Self: 'async_trait,

TotalEscrowForDenom returns the total amount of tokens in escrow based on the denom.